chore: seed data for it tests

+++ src/main/java/dev/jpitarch/ctrlgym/core/models/GymBranchHeartbeatMO.java
@@ -49,7 +49,7 @@ public class GymBranchHeartbeatMO {
 
   @Column(name = "cpu_percent")
   @JsonProperty("cpu_percent")
-  private Double cpuPercent;
+  private BigDecimal cpuPercent;
 
   @Column(name = "memory_percent")
   @JsonProperty("memory_percent")
